- new AA implementation
- uses open addressing with quadratic probing (triangular numbers) and pow2 table
- uses NO_SCAN for entries when applicable
- minimizes alignment gap for values
- calls postblit on aa.keys and aa.values

[Issue 14385 – AA should use open addressing hash](https://issues.dlang.org/show_bug.cgi?id=14385)