@MartinNowak 2015-04-21
[![chart](https://cloud.githubusercontent.com/assets/288976/7243965/e0fc6e20-e7d4-11e4-9294-f8ac1e2cd3af.png)](https://dawg.eu/open_addressing.html)

The conmsg benchmark suffers from increased congestion on the GC lock, because I removed the small inplace AA buckets which take quite a lot of extra memory even though they aren't used for bigger AAs. The congestion problem will get fixed by introducing GC thread-caches.