@rainers 2015-04-21
Looks good to a first cursory inspection.

Can we add some hint to a debugger to figure out what version of the AA implementation is used? Both cv2pdb and mago rebuild the internals to display AA elements.