@MartinNowak 2015-04-21
> Can we add some hint to a debugger to figure out what version of the AA implementation is used?

What kind of hint would work?