@rainers 2015-04-21
> What kind of hint would work?

I think a global version variable should work, as globals are usually not split into seperate COMDATs, so it would not be stripped by the linker:
```
immutable int AA_version = 1;
```
I just tried this, it is built into the same section as the module info.