@MartinNowak 2015-04-21
> immutable int AA_version = 1;

Done as `extern(C) immutable int _aaVersion = 1;`.