@rainers 2015-04-22
Should we move the keys into their own array in lock-step with the bucket array? That would avoid the tiEntry machinery which gets even worse when trying to add RTInfo for a precise GC.

Pro:
- It might need less memory without padding for value alignment. 
- A lot of indirections could be avoided during lookups

Con:
- It could eat more memory for larger key types
- Destructors will run on unused keys that are in their init state.