@MartinNowak 2015-04-22
> Should we move the keys into their own array in lock-step with the bucket array?

Will try if that helps, though I'd like to avoid further optimization discussions as we can always make things faster tomorrow. 

> I think this AA implementation is a nice improvement and fixes a number of issues regarding postblit and destructor calls. It is not a small change, though, so I'd like some AA experts to chime in. 

The implementation is state of the art for hash tables, see https://google-sparsehash.googlecode.com/svn/trunk/doc/implementation.html or http://llvm.org/docs/doxygen/html/DenseMap_8h_source.html.

The main intention here is to improve the old implementation so that we have a solid AA while transitioning to a library implementation. Only that allows to get a fast AA.