@rainers 2015-04-24
> The implementation is state of the art for hash tables

Then I'm slightly disappointed by the limited improvement that the benchmarks show.

It seems there are no objections otherwise, though, so let's get this in. The omission of the `next` field in `Entry` will probably help the precise GC in the benchmarks, too. ;-)