@rainers 2015-04-24
Auto-merge toggled on