@MartinNowak 2015-04-24
> Then I'm slightly disappointed by the limited improvement that the benchmarks show.

Yeah, I was a little disappointed as well, the effect is a bit bigger than it appears b/c only a fraction of the benchmark time is actually spent in the AA. Just added another benchmark #1230.

Lots of more improvements are possible with a library AA.
I have the idea to add one as core.aa that's compatible to the builtin one. Let's see whether that works out.