@MartinNowak 2015-04-25
Particularly testgc3 and bulk cause different GC pool allocations, causing a big part of the difference.

Then I'm also seeing a slight increase in stalled frontend cycles, the mix function seems to be responsible for that.
https://github.com/D-Programming-Language/druntime/pull/1229/files#diff-fdc0da51523ff831dd6cbe33a5bb8b4cR294
I'd very much like to use some bitshift mix instead of a multiplication but I couldn't find one that's good enough. We could move this to TypeInfo_int/ptr though, because strings and structs are well hashed already, and also doing some more work in those virtual functions helps to fill the pipeline.