@rainers 2015-04-26
> Particularly testgc3 and bulk cause different GC pool allocations, causing a big part of the difference.

I expected the new version to use less memory for 64-bit, but it seems not to change: both versions of testgc3 need 247 MB as reported by the GC, though that's probably much more than the actual live memory.
The old implementation had some bad alignment requirements that caused each entry to allocate 64 byte. This should be down to 16 now, but every entry has at least one 16 byte entry in the bucket array.

The 32-bit versions both show 117 MB memory usage, and no big change in garbage collection time.

> Then I'm also seeing a slight increase in stalled frontend cycles, the mix function seems to be responsible for that.

I tried to replace the mix function with a noop, but it did not have a big effect. Not sure how much the testgc3 requires a good hash, though.

I also noticed your replacement of `rep stosb` (it seems very slow on my mobile system, too. According to Agner Fog, these operations have a quite large setup time and only pay off for large block operations). Replacing the memset and the memcpy with a inlinable version tweaked for short sizes improves the benchmark by 5-10%.