@MartinNowak 2015-04-26
> Replacing the memset and the memcpy with a inlinable version tweaked for short sizes improves the benchmark by 5-10%.

Might make sense for <4 bytes, but only if inlineable, because memset does the same check.
[Issue 14458 – very slow ubyte[] assignment (dmd doesn't use memset)](https://issues.dlang.org/show_bug.cgi?id=14458)