@rainers 2015-05-02
I recently tried a few modifications, too. The new implementation grows 3 times for the 200 entries, while the old one did this only once. Starting with an initial bucket size of 64 causes also a single rehashing, but it did not make a large difference.

> Will try to split the bucket array into an array for hashes and one for the pointers. Which allows to save 25% mem on x64 by using only a 32-bit hash but might incur a 2nd cache miss.

I thought about that, too, but did not try it. Event with size_t hashes, it could avoid an additional cache line read if the hash does not match.